Revert "Revert "Fold Channel and CompletionQueue from grpc_impl to grpc""
