+++ include/grpcpp/generic/generic_stub.h
@@ -29,10 +29,9 @@
 #include <grpcpp/support/client_callback_impl.h>
 #include <grpcpp/support/status.h>
 
-namespace grpc_impl {
-class CompletionQueue;
-}  // namespace grpc_impl
 namespace grpc {
+
+class CompletionQueue;
 
 typedef ::grpc_impl::ClientAsyncReaderWriter<ByteBuffer, ByteBuffer>
     GenericClientAsyncReaderWriter;
@@ -55,7 +54,7 @@ class TemplatedGenericStub final {
   std::unique_ptr<
       ::grpc_impl::ClientAsyncReaderWriter<RequestType, ResponseType>>
   PrepareCall(ClientContext* context, const std::string& method,
-              ::grpc_impl::CompletionQueue* cq) {
+              ::grpc::CompletionQueue* cq) {
     return CallInternal(channel_.get(), context, method, cq, false, nullptr);
   }
 
@@ -65,8 +64,7 @@ class TemplatedGenericStub final {
   /// succeeded (i.e. the call won't proceed if the return value is nullptr).
   std::unique_ptr<::grpc_impl::ClientAsyncResponseReader<ResponseType>>
   PrepareUnaryCall(ClientContext* context, const std::string& method,
-                   const RequestType& request,
-                   ::grpc_impl::CompletionQueue* cq) {
+                   const RequestType& request, ::grpc::CompletionQueue* cq) {
     return std::unique_ptr<
         ::grpc_impl::ClientAsyncResponseReader<ResponseType>>(
         grpc_impl::internal::ClientAsyncResponseReaderFactory<
@@ -86,7 +84,7 @@ class TemplatedGenericStub final {
   std::unique_ptr<
       ::grpc_impl::ClientAsyncReaderWriter<RequestType, ResponseType>>
   Call(ClientContext* context, const std::string& method,
-       ::grpc_impl::CompletionQueue* cq, void* tag) {
+       ::grpc::CompletionQueue* cq, void* tag) {
     return CallInternal(channel_.get(), context, method, cq, true, tag);
   }
 
@@ -205,7 +203,7 @@ class TemplatedGenericStub final {
   std::unique_ptr<
       ::grpc_impl::ClientAsyncReaderWriter<RequestType, ResponseType>>
   CallInternal(grpc::ChannelInterface* channel, ClientContext* context,
-               const std::string& method, ::grpc_impl::CompletionQueue* cq,
+               const std::string& method, ::grpc::CompletionQueue* cq,
                bool start, void* tag) {
     return std::unique_ptr<
         ::grpc_impl::ClientAsyncReaderWriter<RequestType, ResponseType>>(
