+++ include/grpcpp/impl/codegen/channel_interface.h
@@ -26,7 +26,6 @@
 
 namespace grpc_impl {
 class ClientContext;
-class CompletionQueue;
 template <class R>
 class ClientReader;
 template <class W>
@@ -56,6 +55,7 @@ class ClientCallbackUnaryFactory;
 
 namespace grpc {
 class ChannelInterface;
+class CompletionQueue;
 
 namespace experimental {
 class DelegatingChannel;
@@ -82,7 +82,7 @@ class ChannelInterface {
   /// deadline expires. \a GetState needs to called to get the current state.
   template <typename T>
   void NotifyOnStateChange(grpc_connectivity_state last_observed, T deadline,
-                           ::grpc_impl::CompletionQueue* cq, void* tag) {
+                           ::grpc::CompletionQueue* cq, void* tag) {
     TimePoint<T> deadline_tp(deadline);
     NotifyOnStateChangeImpl(last_observed, deadline_tp.raw_time(), cq, tag);
   }
@@ -136,13 +136,13 @@ class ChannelInterface {
   friend class ::grpc::internal::InterceptedChannel;
   virtual internal::Call CreateCall(const internal::RpcMethod& method,
                                     ::grpc_impl::ClientContext* context,
-                                    ::grpc_impl::CompletionQueue* cq) = 0;
+                                    ::grpc::CompletionQueue* cq) = 0;
   virtual void PerformOpsOnCall(internal::CallOpSetInterface* ops,
                                 internal::Call* call) = 0;
   virtual void* RegisterMethod(const char* method) = 0;
   virtual void NotifyOnStateChangeImpl(grpc_connectivity_state last_observed,
                                        gpr_timespec deadline,
-                                       ::grpc_impl::CompletionQueue* cq,
+                                       ::grpc::CompletionQueue* cq,
                                        void* tag) = 0;
   virtual bool WaitForStateChangeImpl(grpc_connectivity_state last_observed,
                                       gpr_timespec deadline) = 0;
@@ -156,8 +156,8 @@ class ChannelInterface {
   // change (even though this is private and non-API)
   virtual internal::Call CreateCallInternal(
       const internal::RpcMethod& /*method*/,
-      ::grpc_impl::ClientContext* /*context*/,
-      ::grpc_impl::CompletionQueue* /*cq*/, size_t /*interceptor_pos*/) {
+      ::grpc_impl::ClientContext* /*context*/, ::grpc::CompletionQueue* /*cq*/,
+      size_t /*interceptor_pos*/) {
     return internal::Call();
   }
 
@@ -169,7 +169,7 @@ class ChannelInterface {
   // Returns nullptr (rather than being pure) since this is a post-1.0 method
   // and adding a new pure method to an interface would be a breaking change
   // (even though this is private and non-API)
-  virtual ::grpc_impl::CompletionQueue* CallbackCQ() { return nullptr; }
+  virtual ::grpc::CompletionQueue* CallbackCQ() { return nullptr; }
 };
 }  // namespace grpc
