+++ include/grpcpp/impl/codegen/intercepted_channel.h
@@ -21,11 +21,8 @@
 
 #include <grpcpp/impl/codegen/channel_interface.h>
 
-namespace grpc_impl {
-class CompletionQueue;
-}
-
 namespace grpc {
+class CompletionQueue;
 
 namespace internal {
 
@@ -50,7 +47,7 @@ class InterceptedChannel : public ChannelInterface {
       : channel_(channel), interceptor_pos_(pos) {}
 
   Call CreateCall(const RpcMethod& method, ::grpc_impl::ClientContext* context,
-                  ::grpc_impl::CompletionQueue* cq) override {
+                  ::grpc::CompletionQueue* cq) override {
     return channel_->CreateCallInternal(method, context, cq, interceptor_pos_);
   }
 
@@ -63,7 +60,7 @@ class InterceptedChannel : public ChannelInterface {
 
   void NotifyOnStateChangeImpl(grpc_connectivity_state last_observed,
                                gpr_timespec deadline,
-                               ::grpc_impl::CompletionQueue* cq,
+                               ::grpc::CompletionQueue* cq,
                                void* tag) override {
     return channel_->NotifyOnStateChangeImpl(last_observed, deadline, cq, tag);
   }
@@ -72,7 +69,7 @@ class InterceptedChannel : public ChannelInterface {
     return channel_->WaitForStateChangeImpl(last_observed, deadline);
   }
 
-  ::grpc_impl::CompletionQueue* CallbackCQ() override {
+  ::grpc::CompletionQueue* CallbackCQ() override {
     return channel_->CallbackCQ();
   }
