+++ src/cpp/server/async_generic_service.cc
@@ -24,8 +24,8 @@ namespace grpc {
 
 void AsyncGenericService::RequestCall(
     GenericServerContext* ctx, GenericServerAsyncReaderWriter* reader_writer,
-    ::grpc_impl::CompletionQueue* call_cq,
-    ::grpc_impl::ServerCompletionQueue* notification_cq, void* tag) {
+    ::grpc::CompletionQueue* call_cq,
+    ::grpc::ServerCompletionQueue* notification_cq, void* tag) {
   server_->RequestAsyncGenericCall(ctx, reader_writer, call_cq, notification_cq,
                                    tag);
 }
