+++ src/cpp/server/server_builder.cc
@@ -68,14 +68,14 @@ ServerBuilder::~ServerBuilder() {
   }
 }
 
-std::unique_ptr<grpc_impl::ServerCompletionQueue>
-ServerBuilder::AddCompletionQueue(bool is_frequently_polled) {
-  grpc_impl::ServerCompletionQueue* cq = new grpc_impl::ServerCompletionQueue(
+std::unique_ptr<grpc::ServerCompletionQueue> ServerBuilder::AddCompletionQueue(
+    bool is_frequently_polled) {
+  grpc::ServerCompletionQueue* cq = new grpc::ServerCompletionQueue(
       GRPC_CQ_NEXT,
       is_frequently_polled ? GRPC_CQ_DEFAULT_POLLING : GRPC_CQ_NON_LISTENING,
       nullptr);
   cqs_.push_back(cq);
-  return std::unique_ptr<grpc_impl::ServerCompletionQueue>(cq);
+  return std::unique_ptr<grpc::ServerCompletionQueue>(cq);
 }
 
 ServerBuilder& ServerBuilder::RegisterService(Service* service) {
@@ -275,11 +275,10 @@ std::unique_ptr<grpc::Server> ServerBuilder::BuildAndStart() {
   // This is different from the completion queues added to the server via
   // ServerBuilder's AddCompletionQueue() method (those completion queues
   // are in 'cqs_' member variable of ServerBuilder object)
-  std::shared_ptr<
-      std::vector<std::unique_ptr<grpc_impl::ServerCompletionQueue>>>
+  std::shared_ptr<std::vector<std::unique_ptr<grpc::ServerCompletionQueue>>>
       sync_server_cqs(
-          std::make_shared<std::vector<
-              std::unique_ptr<grpc_impl::ServerCompletionQueue>>>());
+          std::make_shared<
+              std::vector<std::unique_ptr<grpc::ServerCompletionQueue>>>());
 
   bool has_frequently_polled_cqs = false;
   for (const auto& cq : cqs_) {
@@ -307,8 +306,8 @@ std::unique_ptr<grpc::Server> ServerBuilder::BuildAndStart() {
 
     // Create completion queues to listen to incoming rpc requests
     for (int i = 0; i < sync_server_settings_.num_cqs; i++) {
-      sync_server_cqs->emplace_back(new grpc_impl::ServerCompletionQueue(
-          GRPC_CQ_NEXT, polling_type, nullptr));
+      sync_server_cqs->emplace_back(
+          new grpc::ServerCompletionQueue(GRPC_CQ_NEXT, polling_type, nullptr));
     }
   }
