+++ test/cpp/util/create_test_channel.h
@@ -26,12 +26,8 @@
 #include <grpcpp/security/credentials.h>
 #include <grpcpp/support/channel_arguments.h>
 
-namespace grpc_impl {
-
-class Channel;
-}
-
 namespace grpc {
+class Channel;
 
 namespace testing {
